+++ src/Testing/FakeOutboundEventPublisher.php
@@ -12,12 +12,19 @@ declare(strict_types=1);
 
 namespace CloudCreativity\Modules\Testing;
 
+use ArrayAccess;
 use CloudCreativity\Modules\Contracts\Application\Ports\Driven\OutboundEventPublisher;
 use CloudCreativity\Modules\Contracts\Toolkit\Messages\IntegrationEvent;
 use Countable;
+use Generator;
+use IteratorAggregate;
 use LogicException;
 
-class FakeOutboundEventPublisher implements OutboundEventPublisher, Countable
+/**
+ * @implements ArrayAccess<int, IntegrationEvent>
+ * @implements IteratorAggregate<int, IntegrationEvent>
+ */
+class FakeOutboundEventPublisher implements OutboundEventPublisher, Countable, ArrayAccess, IteratorAggregate
 {
     /**
      * @var list<IntegrationEvent>
@@ -27,6 +34,34 @@ class FakeOutboundEventPublisher implements OutboundEventPublisher, Countable
     public function publish(IntegrationEvent $event): void
     {
         $this->events[] = $event;
+    }
+
+    public function offsetExists(mixed $offset): bool
+    {
+        return isset($this->events[$offset]);
+    }
+
+    public function offsetGet(mixed $offset): IntegrationEvent
+    {
+        return $this->events[$offset];
+    }
+
+    public function offsetSet(mixed $offset, mixed $value): void
+    {
+        throw new LogicException('Cannot set integration events.');
+    }
+
+    public function offsetUnset(mixed $offset): void
+    {
+        throw new LogicException('Cannot unset integration events.');
+    }
+
+    /**
+     * @return Generator<int, IntegrationEvent>
+     */
+    public function getIterator(): Generator
+    {
+        yield from $this->events;
     }
 
     public function count(): int