+++ src/Testing/FakeDomainEventDispatcher.php
@@ -12,12 +12,19 @@ declare(strict_types=1);
 
 namespace CloudCreativity\Modules\Testing;
 
+use ArrayAccess;
 use CloudCreativity\Modules\Contracts\Domain\Events\DomainEvent;
 use CloudCreativity\Modules\Contracts\Domain\Events\DomainEventDispatcher;
 use Countable;
+use Generator;
+use IteratorAggregate;
 use LogicException;
 
-class FakeDomainEventDispatcher implements DomainEventDispatcher, Countable
+/**
+ * @implements ArrayAccess<int, DomainEvent>
+ * @implements IteratorAggregate<int, DomainEvent>
+ */
+class FakeDomainEventDispatcher implements DomainEventDispatcher, Countable, ArrayAccess, IteratorAggregate
 {
     /**
      * @var list<DomainEvent>
@@ -27,6 +34,34 @@ class FakeDomainEventDispatcher implements DomainEventDispatcher, Countable
     public function dispatch(DomainEvent $event): void
     {
         $this->events[] = $event;
+    }
+
+    public function offsetExists(mixed $offset): bool
+    {
+        return isset($this->events[$offset]);
+    }
+
+    public function offsetGet(mixed $offset): DomainEvent
+    {
+        return $this->events[$offset];
+    }
+
+    public function offsetSet(mixed $offset, mixed $value): void
+    {
+        throw new LogicException('Cannot set domain events.');
+    }
+
+    public function offsetUnset(mixed $offset): void
+    {
+        throw new LogicException('Cannot unset domain events.');
+    }
+
+    /**
+     * @return Generator<int, DomainEvent>
+     */
+    public function getIterator(): Generator
+    {
+        yield from $this->events;
     }
 
     public function count(): int