+++ src/Testing/FakeQueue.php
@@ -12,12 +12,19 @@ declare(strict_types=1);
 
 namespace CloudCreativity\Modules\Testing;
 
+use ArrayAccess;
 use CloudCreativity\Modules\Contracts\Application\Ports\Driven\Queue;
 use CloudCreativity\Modules\Contracts\Toolkit\Messages\Command;
 use Countable;
+use Generator;
+use IteratorAggregate;
 use LogicException;
 
-class FakeQueue implements Queue, Countable
+/**
+ * @implements ArrayAccess<int, Command>
+ * @implements IteratorAggregate<int, Command>
+ */
+class FakeQueue implements Queue, Countable, ArrayAccess, IteratorAggregate
 {
     /**
      * @var list<Command>
@@ -27,6 +34,34 @@ class FakeQueue implements Queue, Countable
     public function push(Command $command): void
     {
         $this->commands[] = $command;
+    }
+
+    public function offsetExists(mixed $offset): bool
+    {
+        return isset($this->commands[$offset]);
+    }
+
+    public function offsetGet(mixed $offset): Command
+    {
+        return $this->commands[$offset];
+    }
+
+    public function offsetSet(mixed $offset, mixed $value): void
+    {
+        throw new LogicException('Cannot set commands.');
+    }
+
+    public function offsetUnset(mixed $offset): void
+    {
+        throw new LogicException('Cannot unset commands.');
+    }
+
+    /**
+     * @return Generator<int, Command>
+     */
+    public function getIterator(): Generator
+    {
+        yield from $this->commands;
     }
 
     public function count(): int